Fix MSVC signed/unsigned comparison warnings in for loops

+++ guetzli/processor.cc
@@ -83,7 +83,7 @@ void RemoveOriginalQuantization(JPEGData* jpg, int q_in[3][kDCTBlockSize]) {
     JPEGComponent& c = jpg->components[i];
     const int* q = &jpg->quant[c.quant_idx].values[0];
     memcpy(&q_in[i][0], q, kDCTBlockSize * sizeof(q[0]));
-    for (int j = 0; j < c.coeffs.size(); ++j) {
+    for (size_t j = 0; j < c.coeffs.size(); ++j) {
       c.coeffs[j] *= q[j % kDCTBlockSize];
     }
   }
@@ -224,7 +224,7 @@ class QuantMatrixGenerator {
       }
       GetQuantMatrixWithHeuristicScore(hscore, q);
       bool retry = false;
-      for (int i = 0; i < quants_.size(); ++i) {
+      for (size_t i = 0; i < quants_.size(); ++i) {
         if (CompareQuantMatrices(&q[0][0], &quants_[i].q[0][0]) == 0) {
           if (quants_[i].dist_ok) {
             hscore_a_ = hscore;
@@ -388,7 +388,7 @@ void Processor::ComputeBlockZeroingOrder(
   while (!input_order.empty()) {
     float best_err = 1e17;
     int best_i = 0;
-    for (int i = 0; i < std::min<size_t>(params_.zeroing_greedy_lookahead,
+    for (size_t i = 0; i < std::min<size_t>(params_.zeroing_greedy_lookahead,
                                          input_order.size());
          ++i) {
       coeff_t candidate_block[kBlockSize];
@@ -435,7 +435,7 @@ void Processor::ComputeBlockZeroingOrder(
     (*output_order)[i].block_err = min_err;
   }
   // Cut off at the block error limit.
-  int num = 0;
+  size_t num = 0;
   while (num < output_order->size() &&
          (*output_order)[num].block_err <= comparator_->BlockErrorLimit()) {
     ++num;
@@ -487,13 +487,13 @@ size_t ComputeEntropyCodes(const std::vector<JpegHistogram>& histograms,
       histograms.size() * JpegHistogram::kSize);
   ClusterHistograms(&clustered[0], &num, &indexes[0], &clustered_depths[0]);
   depths->resize(clustered_depths.size());
-  for (int i = 0; i < histograms.size(); ++i) {
+  for (size_t i = 0; i < histograms.size(); ++i) {
     memcpy(&(*depths)[i * JpegHistogram::kSize],
            &clustered_depths[indexes[i] * JpegHistogram::kSize],
            JpegHistogram::kSize);
   }
   size_t histogram_size = 0;
-  for (int i = 0; i < num; ++i) {
+  for (size_t i = 0; i < num; ++i) {
     histogram_size += HistogramHeaderCost(clustered[i]) / 8;
   }
   return histogram_size;
@@ -502,7 +502,7 @@ size_t ComputeEntropyCodes(const std::vector<JpegHistogram>& histograms,
 size_t EntropyCodedDataSize(const std::vector<JpegHistogram>& histograms,
                             const std::vector<uint8_t>& depths) {
   size_t numbits = 0;
-  for (int i = 0; i < histograms.size(); ++i) {
+  for (size_t i = 0; i < histograms.size(); ++i) {
     numbits += HistogramEntropyCost(
         histograms[i], &depths[i * JpegHistogram::kSize]);
   }
@@ -605,7 +605,7 @@ void Processor::SelectFrequencyMasking(const JPEGData& jpg, OutputImage* img,
               continue;
             }
             if (direction > 0) {
-              for (int i = last_index; i < order.size(); ++i) {
+              for (size_t i = last_index; i < order.size(); ++i) {
                 float val = ((order[i].block_err - max_err) /
                              block_weight[block_ix]);
                 global_order.push_back(std::make_pair(block_ix, val));
@@ -641,7 +641,7 @@ void Processor::SelectFrequencyMasking(const JPEGData& jpg, OutputImage* img,
       if (direction > 0 && comparator_->DistanceOK(1.0)) {
         rel_size_delta = 0.05;
       }
-      size_t min_size_delta = base_size * rel_size_delta;
+      double min_size_delta = base_size * rel_size_delta;
 
       float coeffs_to_change_per_block =
           direction > 0 ? 2.0 : factor_x * factor_y * 0.2;
@@ -661,7 +661,7 @@ void Processor::SelectFrequencyMasking(const JPEGData& jpg, OutputImage* img,
       float val_threshold = 0.0;
       int changed_coeffs = 0;
       int est_jpg_size = prev_size;
-      for (int i = 0; i < global_order.size(); ++i) {
+      for (size_t i = 0; i < global_order.size(); ++i) {
         const int block_ix = global_order[i].first;
         const int block_x = block_ix % block_width;
         const int block_y = block_ix / block_width;
@@ -760,7 +760,7 @@ bool Processor::ProcessJpegData(const Params& params, const JPEGData& jpg_in,
     input_is_420 = true;
   } else {
     fprintf(stderr, "Unsupported sampling factors:");
-    for (int i = 0; i < jpg_in.components.size(); ++i) {
+    for (size_t i = 0; i < jpg_in.components.size(); ++i) {
       fprintf(stderr, " %dx%d", jpg_in.components[i].h_samp_factor,
               jpg_in.components[i].v_samp_factor);
     }
